% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMATT.R
\name{CMATT}
\alias{CMATT}
\title{This function changes a 3D array to a matrix using row-bind}
\usage{
CMATT(darr, nobj)
}
\arguments{
\item{darr}{3D array of dimension =(p,d,n)}

\item{nobj}{Number of object (=n)}
}
\value{
matrix of dimension (pn*d)
}
\description{
This function changes a 3D array to a matrix using row-bind
}
\examples{
a =  sample(1:5,24,replace = TRUE)
dim(a) = c(3,2,4)
print("3-dimension array:")
print(a)
b=CMAT(a,nobj=4)
print(b)
}
\keyword{CMATT}
