% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmccode_2D_A.R
\name{mcmccode_2D_A}
\alias{mcmccode_2D_A}
\title{function for MCMC sampling for 2D landmark data (Gaussian likelihood with Anisotropic Error Variance)}
\usage{
mcmccode_2D_A(myData, choice, Nlandmark, tune, Nsample)
}
\arguments{
\item{myData}{3D array containing 2 dimensional landmark}

\item{choice}{Which to compare with 1st object or mean obj}

\item{Nlandmark}{How many landmarks are there for each object? (usually nrow(myData\link{,,1}))}

\item{tune}{Tuning value of MCMC sampler}

\item{Nsample}{Number of MCMC sample desired}
}
\value{
matrix containing samples from posterior density of parameter
}
\description{
Draws posterior from 8 parameters "c1","c2", "b", "theta", "Sigma_0", "Sigma_1", "Sigma_2", "delta"
}
\details{
"c1","c2" = Location parameter
#' "b" = dilation parameter
"theta" =Rotation angle
"Sigma_k" = Anisotropic error variation parameter (k=0,1,2)
Note that, here we are assuming Isotropy of error variance
}
\examples{
\dontrun{
require(shapes)
data(apes)
myData = apes$x
ape_ress_10000=mcmccode_2D_A(apes$x,10,nrow(myData[,,1]), tune=1,10)
}

}
\keyword{mcmccode_2D_A}
