% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_avl.R
\name{Test_avl}
\alias{Test_avl}
\title{An remodified Available Function for testing the R Package name}
\usage{
Test_avl(variable)
}
\arguments{
\item{variable}{The package name you want to check whether available or not. It uses the function (available) from the R package "available"}
}
\value{
Valid/Invalid List
}
\description{
This function allows us to express availability of names, validity, whether Available on CRAN, whether Available on Bioconductor, whether Available on GitHub, etc.
}
\examples{
Test_avl("a")
Test_avl("ABCD")
Test_avl("MASS")
}
\keyword{Test_avl}
